Revert questionable conversion
