+++ quadratic-core/src/test_util.rs
@@ -1,31 +1,14 @@
-use std::str::FromStr;
-
 use crate::{
     controller::GridController,
     grid::{Bold, SheetId},
     CellValue, Pos, Rect,
 };
 
-use bigdecimal::BigDecimal;
 use tabled::{
     builder::Builder,
     settings::Color,
     settings::{Modify, Style},
 };
-
-fn to_cell_value(value: &str) -> CellValue {
-    let parsed = CellValue::strip_percentage(CellValue::strip_currency(value)).trim();
-    let number = BigDecimal::from_str(parsed);
-    let is_true = parsed.eq_ignore_ascii_case("true");
-    let is_false = parsed.eq_ignore_ascii_case("false");
-    let is_bool = is_true || is_false;
-
-    match (number, is_bool) {
-        (Ok(number), false) => CellValue::Number(number),
-        (_, true) => CellValue::Logical(is_true),
-        _ => CellValue::Text(String::from(value)),
-    }
-}
 
 /// Run an assertion that a cell value is equal to the given value
 pub fn assert_cell_value(
@@ -42,7 +25,7 @@ pub fn assert_cell_value(
     let expected = if value == "" {
         CellValue::Blank
     } else {
-        to_cell_value(value)
+        CellValue::to_cell_value(value)
     };
 
     assert_eq!(
